[refactor predominant bp algorithm] refactor algo
remove 3 old functions to introduce new function
