+++ hypertension/dc-7101-algorithms/py-root/lib/algorithms/bp_sufficiency.py
@@ -1,12 +1,109 @@
 from datetime import datetime
 from dateutil.relativedelta import relativedelta
+import operator
 
-from .utils import (
-    tally_diastolic_counts,
-    tally_systolic_counts,
-    calculate_reading_from_buckets,
-    bp_readings_meet_date_specs
-)
+def bp_readings_meet_date_specs(date_of_claim, bp_readings):
+    """
+    Determine from a list of BP readings if there exists two readings with a 1 month and 6 month date window 
+
+    :param date_of_claim: string of date of claim
+    :type date_of_claim: string
+    :param bp_readings: list of blood pressure readings
+    :type bp_readings: list
+    :return: boolean indicating if readings within the two date windows are present
+    :rtype: boolean
+    """
+
+    reading_within_one_month = False
+    reading_within_six_months = False
+    date_of_claim_date = datetime.strptime(date_of_claim, "%Y-%m-%d").date()
+
+    for reading in bp_readings:
+        bp_reading_date = datetime.strptime(reading["date"], "%Y-%m-%d").date()
+        one_month_difference = (date_of_claim_date - bp_reading_date).days <= 30
+        six_month_difference = (date_of_claim_date - bp_reading_date).days <= 180
+
+        if (one_month_difference and not reading_within_one_month):
+            reading_within_one_month = True
+        elif (one_month_difference and reading_within_one_month):
+            reading_within_six_months = True
+        elif (six_month_difference and not reading_within_six_months):
+            reading_within_six_months = True
+
+    return reading_within_one_month and reading_within_six_months
+
+def calculate_predominant_ratings(bp_readings):
+    """
+    Calculate the predominant diastolic and systolic values from a list of BP readings
+
+    :param bp_readings: list of blood pressure readings
+    :type bp_readings: list
+    :return: dictionary with systolic and diastolic values 
+    :rtype: dict
+    """
+    diastolic_130_and_above = []
+    diastolic_120_to_129 = []
+    diastolic_110_to_119 = []
+    diastolic_100_to_109 = []
+    diastolic_0_to_99 = []
+
+    systolic_200_and_above = []
+    systolic_160_to_199 = []
+    systolic_0_to_159 = []
+
+    for reading in bp_readings:
+        diastolic_value = reading["diastolic"]
+        systolic_value = reading["systolic"]
+
+        if diastolic_value >= 130:
+            diastolic_130_and_above.append(reading)
+        elif diastolic_value >= 120 and diastolic_value < 130:
+            diastolic_120_to_129.append(reading)
+        elif diastolic_value >= 110 and diastolic_value < 120:
+            diastolic_110_to_119.append(reading)
+        elif diastolic_value >= 100 and diastolic_value < 110:
+            diastolic_100_to_109.append(reading)
+        else:
+            diastolic_0_to_99.append(reading)
+
+        if systolic_value >= 200:
+            systolic_200_and_above.append(reading)
+        elif systolic_value >= 160:
+            systolic_160_to_199.append(reading)
+        else:
+            systolic_0_to_159.append(reading)
+
+    list_of_diastolic_lists = [
+        diastolic_120_to_129,
+        diastolic_110_to_119,
+        diastolic_100_to_109,
+        diastolic_0_to_99
+    ]
+
+    list_of_systolic_lists =[
+        systolic_160_to_199,
+        systolic_0_to_159
+    ]
+
+    longest_diastolic_list = diastolic_130_and_above
+    longest_systolic_list = systolic_200_and_above
+
+    for diastolic_list in list_of_diastolic_lists:
+        if len(diastolic_list) > len(longest_diastolic_list):
+            longest_diastolic_list = diastolic_list
+
+    for systolic_list in list_of_systolic_lists:
+        if len(systolic_list) > len(longest_systolic_list):
+            longest_systolic_list = systolic_list
+
+    diastolic_list.sort(key=operator.itemgetter("date"))
+
+    systolic_list.sort(key=operator.itemgetter("date"))
+
+    return {
+        "systolic_value": longest_systolic_list[-1]["systolic"],
+        "diastolic_value": longest_diastolic_list[-1]["diastolic"]
+    }
 
 def sufficient_to_autopopulate (request_body):
     """
@@ -45,9 +142,8 @@ def sufficient_to_autopopulate (request_body):
             predominance_calculation["predominant_systolic_reading"] = first_reading["systolic"] if first_reading["systolic"] > second_reading["systolic"] else second_reading["systolic"] 
 
         elif len(valid_bp_readings) > 2:
-            bucketed_diastolic_readings = tally_diastolic_counts(valid_bp_readings)
-            predominance_calculation["predominant_diastolic_reading"] = calculate_reading_from_buckets(bucketed_diastolic_readings, valid_bp_readings, True)
-            bucketed_systolic_readings = tally_systolic_counts(valid_bp_readings)
-            predominance_calculation["predominant_systolic_reading"] = calculate_reading_from_buckets(bucketed_systolic_readings, valid_bp_readings, False)
+            results = calculate_predominant_ratings(valid_bp_readings)
+            predominance_calculation["predominant_diastolic_reading"] = results["diastolic_value"]
+            predominance_calculation["predominant_systolic_reading"] = results["systolic_value"]
 
     return predominance_calculation
